add copying data

+++ test_scripts/test_release_data.py
@@ -13,7 +13,9 @@ import shutil
 if __name__ == '__main__':
     origin_root = '/scratches/flyer_2/fx221/localization/outputs'
     dataset_root = '/scratches/flyer_3/fx221/dataset'
-    save_root = '/scratches/flyer_2/fx221/publications/pram_data/3D-models'
+    landmark_root = '/scratches/flyer_3/fx221/exp/localizer/resnet4x-20230511-210205-pho-0005-gm'
+    # save_root = '/scratches/flyer_2/fx221/publications/pram_data/3D-models'
+    save_root = '/scratches/flyer_2/fx221/publications/pram_data/landmarks'
     # dataset = '7Scenes'
     # dataset = '12Scenes'
     # dataset = 'CambridgeLandmarks'
@@ -22,12 +24,39 @@ if __name__ == '__main__':
 
     with open(config_file, 'rt') as f:
         configs = yaml.load(f, Loader=yaml.Loader)
-    print(configs)
+    # print(configs)
     all_scenes = configs['scenes']
     for scene in all_scenes:
-        scene_path = osp.join(dataset_root, dataset, scene)
+        scene_path = osp.join(landmark_root, dataset, scene)
         target_scene_path = osp.join(save_root, dataset, scene)
+        os.makedirs(target_scene_path, exist_ok=True)
 
+        n_cluster = configs[scene]['n_cluster']
+        cluster_mode = configs[scene]['cluster_mode']
+        cluster_method = configs[scene]['cluster_method']
+
+        file_names = [
+            'compress_model_birch',
+            'point3D_desc.npy',
+            'point3D_cluster_n{:d}_{:s}_birch.npy'.format(n_cluster, cluster_mode),
+            'point3D_vrf_n{:d}_{:s}_birch.npy'.format(n_cluster, cluster_mode),
+        ]
+
+        for file_name in file_names:
+            source_path = osp.join(scene_path, file_name)
+            is_a_file = osp.isfile(source_path)
+            is_a_dir = osp.isdir(source_path)
+            if not is_a_file and not is_a_dir:
+                print('{:s} not exist'.format(source_path))
+                continue
+
+            if is_a_file:
+                shutil.copy2(source_path, osp.join(target_scene_path, file_name))
+            else:
+                shutil.copytree(source_path, osp.join(target_scene_path, file_name))
+
+        ''' copying files for sfm '''
+        '''
         file_names = [
             'pairs-db-covis20.txt',
             'pairs-query-netvlad20.txt',
@@ -35,9 +64,14 @@ if __name__ == '__main__':
             'queries_with_intrinsics.txt',
         ]
         for file_name in file_names:
+            if not osp.isfile(osp.join(scene_path, file_name)):
+                print('{:s} not exist'.format(osp.join(scene_path, file_name)))
+                continue
             shutil.copy2(osp.join(scene_path, file_name), osp.join(target_scene_path, file_name))
+        '''
 
         ''' copying 3D models'''
+        '''
         sfm_path = 'sfm_resnet4x-20230511-210205-pho-0005-gm'
         raw_sfm_path = osp.join(origin_root, dataset, scene, sfm_path)
 
@@ -48,3 +82,4 @@ if __name__ == '__main__':
         shutil.copy2(osp.join(raw_sfm_path, 'images.bin'), osp.join(save_path, 'images.bin'))
         shutil.copy2(osp.join(raw_sfm_path, 'points3D.bin'), osp.join(save_path, 'points3D.bin'))
         shutil.copy2(osp.join(raw_sfm_path, 'statics.txt'), osp.join(save_path, 'statics.txt'))
+        '''
